services/text-service/requirements.txt
anthropic>=0.40.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.57.0
grpcio-tools==1.57.0
pydantic==2.5.0